restructured api folder
